Create CTEs.sql
Query 1: A Common Table Expression (CTE) named top_5_customers is created to find the top 5 customers based on their payment amounts in specific cities. This CTE is then used in the main query to count the number of these top customers in each country. The LIMIT 5 in the CTE ensures that only the top 5 customers are considered, providing valuable insights into customer distribution within countries.